% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_functions.R
\name{char0toNULL}
\alias{char0toNULL}
\title{Convert character(0) or list() into NULL}
\usage{
char0toNULL(x)
}
\arguments{
\item{x}{A character}
}
\value{
x or NULL
}
\description{
Auxiliary internal function.
}
