% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_sequence.R
\name{OEIS_bfile}
\alias{OEIS_bfile}
\title{OEIS_bfile: Class constructor}
\usage{
OEIS_bfile(ID)
}
\arguments{
\item{ID}{A string with the OEIS sequence ID}
}
\value{
An object of the class \code{OEIS_bfile}
}
\description{
OEIS_bfile: Class constructor
}
