% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_sequence.R
\name{OEIS_bfile_url}
\alias{OEIS_bfile_url}
\title{OEIS_bfile_url: Get bfile url}
\usage{
OEIS_bfile_url(ID, URL = FALSE)
}
\arguments{
\item{ID}{A string with the OEIS sequence ID}

\item{URL}{A logical that selects if the output is the complete url or
just the bfile file name.}
}
\value{
A string with the sequence bfile url of the bfile file name
}
\description{
OEIS_bfile_url: Get bfile url
}
\examples{
id <- "A000056"
OEIS_bfile_url(id, FALSE)
OEIS_bfile_url(id, TRUE)
}
