% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_sequence.R
\name{OEIS_cf}
\alias{OEIS_cf}
\title{OEIS sequence linked cross references from \code{seq_df} data.frame}
\usage{
OEIS_cf(seq_df)
}
\arguments{
\item{seq_df}{A data.frame with sequence content}
}
\value{
A character vector with OEIS sequences linked cross references IDs
}
\description{
OEIS sequence linked cross references from \code{seq_df} data.frame
}
\examples{
id <- "A240106"
test_seq_html <- OEIS_xml2(id)
seq_df <- OEIS_df(test_seq_html)
OEIS_cf(seq_df)
}
