% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_sequence.R
\name{OEIS_check}
\alias{OEIS_check}
\title{OEIS_check: OEIS ID Validation}
\usage{
OEIS_check(ID)
}
\arguments{
\item{ID}{A string with the OEIS sequence ID}
}
\value{
NULL or an error if ID is not valid
}
\description{
OEIS_check: OEIS ID Validation
}
\examples{
OEIS_check("A002000")
}
