% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_date.R
\name{OEIS_date}
\alias{OEIS_date}
\title{OEIS sequence creation date from \code{data.frame}}
\usage{
OEIS_date(x)
}
\arguments{
\item{x}{Can be one of the following:
\itemize{
\item A character string with sequence \code{ID}.
\item A \code{"OEIS_internal"} S3 class object, with sequence internal format.
\item A S3 classes object \code{"OEIS_xml"} and \code{"xml_document"}, with
sequence content from OEIS web.
\item An object of \code{"OEIS_sequence"} class, that contains all sequence
related data.
}}
}
\value{
A \code{Date} object with the OEIS sequence creation date, or
  \code{NULL} if no date is found.
}
\description{
Sequence creation date.
}
\examples{
\dontrun{
id <- "A221863"
OEIS_date(id)
}
}
