% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_description.R
\name{OEIS_description}
\alias{OEIS_description}
\title{Get OEIS sequence description from XML document}
\usage{
OEIS_description(seq_xml)
}
\arguments{
\item{seq_xml}{A xml_document from an OEIS sequence}
}
\value{
A string with the OEIS sequence description
}
\description{
Sequence description gives a brief definition of the sequence. In the
description, a(n) usually denotes the n-th term of the sequence, and n is a
typical subscript. In some cases however n denotes a typical term in the
sequence.
}
\examples{
id <- "A000056"
test_seq_xml <- OEIS_xml2(id)
OEIS_description(test_seq_xml)
}
\seealso{
\code{\link{OEIS_xml2}}
}
