% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_sequence.R
\name{OEIS_df}
\alias{OEIS_df}
\title{Get OEIS sequence terms from xml2 data}
\usage{
OEIS_df(seq_xml)
}
\arguments{
\item{seq_xml}{A xml_document from an OEIS sequence}
}
\value{
A data.frame with sequence content
}
\description{
Get OEIS sequence terms from xml2 data
}
