% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_example.R
\name{OEIS_example}
\alias{OEIS_example}
\title{OEIS sequence example from \code{seq_df} \code{data.frame}}
\usage{
OEIS_example(seq_df)
}
\arguments{
\item{seq_df}{A \code{data.frame} with sequence content.}
}
\value{
A character string with the OEIS sequence examples or \code{NULL} if
  there are no examples.
}
\description{
These lines give expanded information or examples to illustrate the initial
terms of the sequence.
}
\examples{
\dontrun{
id <- "A000108"
test_seq_html <- OEIS_xml2(id)
seq_df <- OEIS_df(test_seq_html)
example <- OEIS_example(seq_df)
cat(example)
}
}
\seealso{
\code{\link{OEIS_df}}

\code{\link{OEIS_xml2}}
}
