% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_formerly.R
\name{OEIS_formerly}
\alias{OEIS_formerly}
\title{Get OEIS sequence former \code{ID} from \code{xml2} data}
\usage{
OEIS_formerly(seq_xml)
}
\arguments{
\item{seq_xml}{A xml_document from an OEIS sequence}
}
\value{
A string with the OEIS former \code{ID} or \code{NULL}, if there is
  no former sequence designation.
}
\description{
This is an alternate \code{ID} for some sequences that also have a 4-digit
M-number, such as M1459, which is the number they carried in "The
Encyclopedia of Integer Sequences" by N.J.A. Sloane and S. Plouffe, Academic
Press, San Diego, CA, 1995. 'Some older sequences also have a 4-digit
N-number, such as N0577, which is the number they carried in the "Handbook of
Integer Sequences", by N. J. A. Sloane, Academic Press, NY, 1973
}
\examples{
id <- "A000056"
test_seq_xml <- OEIS_xml2(id)
OEIS_formerly(test_seq_xml)
}
\seealso{
\code{\link{OEIS_description}}

\code{\link{OEIS_xml2}}
}
