% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_formula.R
\name{OEIS_formula}
\alias{OEIS_formula}
\title{OEIS sequence formula from \code{seq_df} \code{data.frame}}
\usage{
OEIS_formula(seq_df)
}
\arguments{
\item{seq_df}{A \code{data.frame} with sequence content.}
}
\value{
A character string with the OEIS sequence formula or \code{NULL} if
  there are no formula.
}
\description{
These lines give formulae, recurrences, generating functions, etc. for the
sequence.
\strong{a(n)} usually denotes the n-th term of the sequence, and \strong{n}
is a typical subscript.
}
\examples{
\dontrun{
id <- "A010108"
test_seq_html <- OEIS_xml2(id)
seq_df <- OEIS_df(test_seq_html)
example <- OEIS_formula(seq_df)
cat(example)
}
}
\seealso{
\code{\link{OEIS_df}}

\code{\link{OEIS_xml2}}

\code{\link{OEIS_example}}
}
