% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_sequence.R
\name{OEIS_keywords}
\alias{OEIS_keywords}
\title{OEIS sequence keywords from \code{xml2} data}
\usage{
OEIS_keywords(seq_xml)
}
\arguments{
\item{seq_xml}{A xml_document from an OEIS sequence}
}
\value{
A character vector with the OEIS sequence keywords
}
\description{
These lines give keywords describing the sequence.
The actual keywords in use can be found at:
\href{https://oeis.org/eishelp2.html}{Explanation of Terms Used}
}
\examples{
id <- "A000012"
test_seq_html <- OEIS_xml2(id)
OEIS_keywords(test_seq_html)
}
