% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_links.R
\name{OEIS_links}
\alias{OEIS_links}
\title{Links from sequence data}
\usage{
OEIS_links(x)
}
\arguments{
\item{x}{Can be one of the following:
\itemize{
\item A character string with sequence \code{ID}.
\item A \code{OEIS_internal} S3 class object, with sequence internal format.
\item A S3 classes object \code{OEIS_xml} and \code{xml_document}, with
sequence content from OEIS web.
\item An object of \code{OEIS_sequence} class, that contains all sequence
related data.
}}
}
\value{
A character string with the OEIS sequence lines of links or
\code{NULL} if there are no links
}
\description{
Links related to the sequence.
}
\examples{
id <- "A105805"
internal_format <- OEIS_internal_format(id)
lnks <- OEIS_links(internal_format)
cat(lnks, sep = "\n")

}
\seealso{
\itemize{
\item \code{\link[=OEIS_description]{OEIS_description()}}
}

\itemize{
\item \code{\link[=OEIS_author]{OEIS_author()}}
}

\itemize{
\item \code{\link[=OEIS_internal_format]{OEIS_internal_format()}}
}

\itemize{
\item \code{\link[=OEIS_formula]{OEIS_formula()}}
}

\itemize{
\item \code{\link[=OEIS_sequence]{OEIS_sequence()}}
}

\itemize{
\item \code{\link[=OEIS_xml]{OEIS_xml()}}
}
}
