% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_mathematica.R
\name{OEIS_mathematica}
\alias{OEIS_mathematica}
\title{OEIS Wolfram Mathematica code from sequence \code{internal_format}}
\usage{
OEIS_mathematica(x)
}
\arguments{
\item{x}{Can be one of the following:
\itemize{
\item A character string with sequence \code{ID}.
\item A \code{OEIS_internal} S3 class object, with sequence internal format.
\item A S3 classes object \code{OEIS_xml} and \code{xml_document}, with
sequence content from OEIS web.
\item An object of \code{OEIS_sequence} class, that contains all sequence
related data.
}}
}
\value{
A character string with the OEIS sequence lines of examples or
\code{NULL} if there are no examples.
}
\description{
These lines give \code{Mathematica} code to produce the sequence.
}
\examples{

id <- "A105805"
internal_format <- OEIS_internal_format(id)
mathematica <- OEIS_mathematica(internal_format)
cat(mathematica, sep = "\n")

}
\seealso{
\itemize{
\item \code{\link[=OEIS_description]{OEIS_description()}}
}

\itemize{
\item \code{\link[=OEIS_author]{OEIS_author()}}
}

\itemize{
\item \code{\link[=OEIS_internal_format]{OEIS_internal_format()}}
}

\itemize{
\item \code{\link[=OEIS_formula]{OEIS_formula()}}
}

\itemize{
\item \code{\link[=OEIS_sequence]{OEIS_sequence()}}
}

\itemize{
\item \code{\link[=OEIS_xml]{OEIS_xml()}}
}
}
