% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_offset.R
\name{OEIS_offset}
\alias{OEIS_offset}
\title{OEIS sequence offset from \code{xml2} data}
\usage{
OEIS_offset(seq_xml)
}
\arguments{
\item{seq_xml}{A xml_document from an OEIS sequence}
}
\value{
A numeric vector with the OEIS sequence distinct offsets
}
\description{
This line usually gives the subscript of the first term in the sequence. If
the sequence gives the decimal expansion of a constant, the offset is the
number of digits before the decimal point. In the OEIS web internal format,
there is a second offset, which says which term (counting from the left, and
starting with 1), first exceeds 1 in absolute value. This is set to 1 if all
the terms are 0 or +-1
}
\examples{
id <- "A000023"
test_seq_html <- OEIS_xml2(id)
OEIS_offset(test_seq_html)
}
\references{
\url{https://oeis.org/eishelp2.html#RS}
}
\seealso{
\code{\link{OEIS_xml2}}
}
