% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_search_url.R
\name{OEIS_search_url}
\alias{OEIS_search_url}
\title{Find url to search in OEIS a given text}
\usage{
OEIS_search_url(txt)
}
\arguments{
\item{txt}{A string with the text to search in OEIS web}
}
\value{
A string with the search url in OEIS
}
\description{
Find url to search in OEIS a given text
}
\examples{
search_txt <- "Numbers n with property that in base 4 representation"
OEIS_search_url(search_txt)
}
\seealso{
\itemize{
\item \code{\link[=OEIS_web_url]{OEIS_web_url()}}
}

\itemize{
\item \code{\link[=OEIS_sequence]{OEIS_sequence()}}
}
}
