% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_crossrefs.R
\name{OEIS_seqs_in_context}
\alias{OEIS_seqs_in_context}
\title{OEIS Sequences in context from from XML document}
\usage{
OEIS_seqs_in_context(seq_xml)
}
\arguments{
\item{seq_xml}{A \code{xml_document} from an OEIS sequence.}
}
\value{
A character vector OEIS sequences in context IDs
}
\description{
This line show the three sequences \code{ID}s immediately before and after
the sequence in the lexicographic listing.
}
\examples{
id <- "A000112"
test_seq_html <- OEIS_xml2(id)
OEIS_seqs_in_context(test_seq_html)
}
\seealso{
\code{\link{OEIS_seqs_adjacent}}

\code{\link{OEIS_crossrefs}}

\code{\link{OEIS_xml2}}

\code{\link{OEIS_df}}
}
