% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_sequence.R
\name{OEIS_sequence}
\alias{OEIS_sequence}
\title{S3 class constructor for \code{OEIS_sequence}}
\usage{
OEIS_sequence(ID)
}
\arguments{
\item{ID}{A string with the OEIS sequence identifier \code{ID} number.
The A-number or sequence \code{ID} is the absolute catalog number of the
sequence. It consists of "A" followed by 6 digits.}
}
\value{
An object of the S3 class \code{OEIS_sequence}.
}
\description{
S3 class constructor for \code{OEIS_sequence}
}
\examples{
id <- "A003456"
A003456 <- OEIS_sequence(id)
class(A003456)
A003456

}
\references{
\href{https://oeis.org/eishelp2.html#RS}{Explanation of Terms Used in Reply From}
}
\seealso{
\itemize{
\item \code{\link[=OEIS_check]{OEIS_check()}}
}

\itemize{
\item \code{\link[=OEIS_url]{OEIS_url()}}
}

\itemize{
\item \code{\link[=OEIS_description]{OEIS_description()}}
}

\itemize{
\item \code{\link[=OEIS_formerly]{OEIS_formerly()}}
}

\itemize{
\item \code{\link[=OEIS_bfile]{OEIS_bfile()}}
}

\itemize{
\item \code{\link[=OEIS_terms]{OEIS_terms()}}
}

\itemize{
\item \code{\link[=OEIS_xml]{OEIS_xml()}}
}

\itemize{
\item \code{\link[=OEIS_internal_format]{OEIS_internal_format()}}
}

\itemize{
\item \code{\link[=OEIS_df]{OEIS_df()}}
}
}
