% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_status.R
\name{OEIS_status}
\alias{OEIS_status}
\title{OEIS sequence status from \code{seq_df} \code{data.frame}}
\usage{
OEIS_status(seq_df)
}
\arguments{
\item{seq_df}{A \code{data.frame} with sequence content.}
}
\value{
A character string with the OEIS sequence status
}
\description{
Status depends on sequence information updates.
}
\note{
status can be: "approved", "editing", "proposed" or "reviewed"
}
\examples{
\dontrun{
id <- "A000108"
test_seq_html <- OEIS_xml2(id)
seq_df <- OEIS_df(test_seq_html)
OEIS_status(seq_df)
}
}
\seealso{
\code{\link{OEIS_df}}

\code{\link{OEIS_xml2}}
}
