% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_sequence.R
\name{OEIS_terms}
\alias{OEIS_terms}
\title{Get OEIS sequence terms from xml2 data}
\usage{
OEIS_terms(seq_xml)
}
\arguments{
\item{seq_xml}{A xml_document from an OEIS sequence}
}
\value{
A character list with the OEIS sequence terms
}
\description{
Get OEIS sequence terms from xml2 data
}
\examples{
id <- "A000056"
test_seq_html <- OEIS_xml2(id)
OEIS_terms(test_seq_html)
}
