% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_total_sequences.R
\name{OEIS_total_sequences}
\alias{OEIS_total_sequences}
\title{Retrieve the total number of sequences in the OEIS database}
\usage{
OEIS_total_sequences()
}
\value{
An integer with the total number of sequences, or \code{NA_integer_}
if the extraction fails or there is no internet connection.
}
\description{
Retrieve the total number of sequences in the OEIS database
}
\examples{
\dontrun{
OEIS_total_sequences()
}
}
\seealso{
\itemize{
\item \code{\link[=OEIS_sequence]{OEIS_sequence()}}
}
}
