% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_sequence.R
\name{OEIS_url}
\alias{OEIS_url}
\title{Get sequence url}
\usage{
OEIS_url(ID)
}
\arguments{
\item{ID}{A string with the OEIS sequence \code{ID} Number
The A-number or sequence \code{ID} is the absolute catalogue number of the
sequence. It consists of A followed by 6 digits}
}
\value{
A string with the sequence url
}
\description{
Get sequence url
}
\examples{
OEIS_url("A000055")
}
