% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_sequence.R
\name{OEIS_xml2}
\alias{OEIS_xml2}
\title{OEIS_xml2: Get sequence content from OEIS web}
\usage{
OEIS_xml2(ID)
}
\arguments{
\item{ID}{A string with the OEIS sequence ID}
}
\value{
\code{xml_document} from OEIS web with the sequence content
}
\description{
OEIS_xml2: Get sequence content from OEIS web
}
\examples{
 id <- "A000156"
 test_seq_html <- OEIS_xml2(id)
}
