% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_sequence.R
\name{plot.OEIS_sequence}
\alias{plot.OEIS_sequence}
\title{Plot class OEIS_Sequence}
\usage{
\method{plot}{OEIS_sequence}(x, ...)
}
\arguments{
\item{x}{An object of the S3 class \code{"OEIS_sequence"}}

\item{...}{other arguments}
}
\value{
A plot of sequence terms
}
\description{
Plot class OEIS_Sequence
}
\examples{
\dontrun{
id <- "A008406"
x <- OEIS_sequence(id)
plot(x)
}
}
