% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_author.R
\name{OEIS_author}
\alias{OEIS_author}
\title{OEIS sequence author from \code{seq_df} \code{data.frame}}
\usage{
OEIS_author(seq_df)
}
\arguments{
\item{seq_df}{A \code{data.frame} with sequence content}
}
\value{
A character vector with the OEIS sequence author
}
\description{
This function gives the name of the person or persons who contributed the
sequence.
}
\examples{
id <- "A000108"
test_seq_html <- OEIS_xml2(id)
seq_df <- OEIS_df(test_seq_html)
OEIS_author(seq_df)
}
\seealso{
\code{\link{OEIS_df}}

\code{\link{OEIS_xml2}}
}
