% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_author.R
\name{OEIS_author}
\alias{OEIS_author}
\title{OEIS sequence authors from \code{internal format}, or from the sequence
\code{ID}.}
\usage{
OEIS_author(x, email = FALSE)
}
\arguments{
\item{x}{Can be one of the following:
\itemize{
\item A character string with sequence \code{ID}.
\item A \code{"OEIS_internal"} S3 class object, with the sequence internal
format.
\item A object of the S3 classes object\code{"OEIS_xml"} and
\code{"xml_document"}, with the sequence content from OEIS web.
\item An object of the \code{"OEIS_sequence"} class, with that contains all
sequence related data.
}}

\item{email}{A logical, by specifying: \code{email = TRUE}, author's email is
returned, if it is present, together with author's name. The default value
of \code{email} is \code{FALSE}.}
}
\value{
A character vector with the OEIS sequence authors and emails
}
\description{
This function gives the name of the person, or persons, who contributed the
sequence and their email, if it's present and selected.
}
\examples{
\dontrun{
id <- "A169890"
OEIS_author(id, email = TRUE)
}
}
\seealso{
\code{\link{OEIS_description}}

\code{\link{OEIS_internal_format}}

\code{\link{OEIS_sequence}}

\code{\link{OEIS_check}}
}
