% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_author.R
\name{OEIS_author}
\alias{OEIS_author}
\title{OEIS sequence authors from sequence\code{internal_format}}
\usage{
OEIS_author(internal_format, email = FALSE)
}
\arguments{
\item{internal_format}{A S3 object of the class \code{OEIS_internal} with the
sequence internal format.}

\item{email}{A logical if yes author email is returned if present.}
}
\value{
A character vector with the OEIS sequence authors and emails
}
\description{
This function gives the name of the person, or persons, who contributed the
sequence and their email if present and selected.
}
\examples{
\dontrun{
id <- "A169890"
internal_format <- OEIS_internal_format(id)
OEIS_author(internal_format, email = TRUE)
}
}
\seealso{
\code{\link{OEIS_description}}

\code{\link{OEIS_internal_format}}
}
