% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_bfile.R
\name{OEIS_bfile}
\alias{OEIS_bfile}
\title{S3 class constructor for \code{bfile} data}
\usage{
OEIS_bfile(ID)
}
\arguments{
\item{ID}{A string with the OEIS sequence identifer \code{ID} number.
The A-number or sequence \code{ID} is the absolute catalogue number of the
sequence.
It consists of "A" followed by 6 digits.}
}
\value{
A S3 object of the class \code{OEIS_bfile}
}
\description{
A \code{bfile} is a text formatted file that provides many more terms of a
sequence than the typical fifty to eighty terms that can be shown in the main
sequence entry.
}
\examples{
id <- "A000056"
OEIS_bfile(id)
}
\seealso{
\code{\link{OEIS_bfile_url}}

\code{\link{OEIS_sequence}}

\code{\link{OEIS_check}}
}
