% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_sequence.R
\name{OEIS_bfile_url}
\alias{OEIS_bfile_url}
\title{Get bfile url}
\usage{
OEIS_bfile_url(ID, URL = FALSE)
}
\arguments{
\item{ID}{A string with the OEIS sequence \code{ID} Number
The A-number or sequence \code{ID} is the absolute catalogue number of the
sequence. It consists of A followed by 6 digits}

\item{URL}{A logical that selects if the output is the complete url, or just
the \code{bfile} file name.}
}
\value{
A string with the sequence \code{bfile} url, or the \code{bfile} file
  name
}
\description{
Get bfile url
}
\examples{
id <- "A000056"
OEIS_bfile_url(id, FALSE)
OEIS_bfile_url(id, TRUE)
}
