% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_bibtex.R
\name{OEIS_bibtex}
\alias{OEIS_bibtex}
\title{Getting BibTeX citation from OEIS sequence}
\usage{
OEIS_bibtex(x, use_tab = TRUE)
}
\arguments{
\item{x}{An object of the \code{OEIS_sequence} class}

\item{use_tab}{A logical if \code{TRUE} adds a tabulator to BibTeX tags}
}
\value{
A character with the BibTeX citation, the function populates a
MISC entry type, for use when nothing else fits, with the following
fields:
\itemize{
\item AUTHOR: The name(s) of the author(s) (in the case of more than one author,
separated by and).
\item TITLE:  The {O}n-{L}ine {E}ncyclopedia of {I}nteger {S}equences.
\item HOWPUBLISHED: Link to sequence in OEIS.
\item YEAR: The year of publication (or, if unpublished, the year of creation).
\item MONTH: The month of publication (or, if unpublished, the month
of creation).
\item NOTE: Sequence description.
}
}
\description{
Getting BibTeX citation from OEIS sequence
}
\note{
Fields tags are separated by commas. The last tag finishs with a comma,
although is not necessarily. The tag's name is not case-sensitive but
caps are used.
}
\examples{
\dontrun{
x <- OEIS_sequence("A170401")
bib <- OEIS_bibtex(x)
cat(bib, sep = "\\n")
}
}
\references{
\itemize{
\item \url{http://www.bibtex.org/Format/}
\item \url{http://psychedelic-geometry.blogspot.com.es/2010/02/bibtex-automatic-oeis-citations.html}
\item \url{https://github.com/Psychedelic-Geometry/OEIS-Mathematica/wiki}
\item \url{https://en.wikipedia.org/wiki/BibTeX}
}
}
