% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_crossrefs.R
\name{OEIS_cf}
\alias{OEIS_cf}
\title{OEIS sequence linked cross references from sequence information
\code{data.frame}}
\usage{
OEIS_cf(seq_df)
}
\arguments{
\item{seq_df}{A \code{data.frame} with sequence content.}
}
\value{
A character vector with OEIS sequence linked cross references
  \code{ID}s
}
\description{
OEIS sequence linked cross references from sequence information
\code{data.frame}
}
\examples{
id <- "A240106"
test_seq_html <- OEIS_xml2(id)
seq_df <- OEIS_df(test_seq_html)
OEIS_cf(seq_df)
}
\seealso{
\code{\link{OEIS_seqs_adjacent}}

\code{\link{OEIS_crossrefs}}

\code{\link{OEIS_xml2}}

\code{\link{OEIS_df}}
}
