% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_check.R
\name{OEIS_check}
\alias{OEIS_check}
\title{OEIS ID Validation}
\usage{
OEIS_check(ID)
}
\arguments{
\item{ID}{A string with the OEIS sequence identifer \code{ID} number.
The A-number or sequence \code{ID} is the absolute catalogue number of the
sequence.
It consists of "A" followed by 6 digits.}
}
\value{
An object of class \code{"OEIS_ID"} or an error if \code{ID} is not
  an OEIS valid ID
}
\description{
OEIS ID Validation
}
\examples{
OEIS_check("A002000")
}
