% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_crossrefs.R
\name{OEIS_crossrefs}
\alias{OEIS_crossrefs}
\title{OEIS Cross references from XML document}
\usage{
OEIS_crossrefs(seq_xml)
}
\arguments{
\item{seq_xml}{A class \code{"xml_document"} from an OEIS sequence.}
}
\value{
An object of the class \code{OEIS_crossrefs} including referenced
  sequences, sequences in context and adjacent sequences.
}
\description{
Cross references IDs
}
\examples{
\dontrun{
id <- "A000112"
seq_html <- OEIS_xml(id)
OEIS_crossrefs(seq_html)
}
}
\seealso{
\code{\link{OEIS_seqs_in_context}}

\code{\link{OEIS_seqs_adjacent}}

\code{\link{OEIS_cf}}

\code{\link{OEIS_df}}

\code{\link{OEIS_xml}}
}
