% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_sequence.R
\name{OEIS_crossrefs}
\alias{OEIS_crossrefs}
\title{OEIS Cross references from \code{xml2} data}
\usage{
OEIS_crossrefs(seq_xml)
}
\arguments{
\item{seq_xml}{A xml_document from an OEIS sequence}
}
\value{
An object of the class \code{OEIS_crossrefs} with sequences in
 context and adjacent sequences
}
\description{
Cross references IDs
}
\examples{
id <- "A000112"
test_seq_html <- OEIS_xml2(id)
OEIS_crossrefs(test_seq_html)
}
\seealso{
\code{\link{OEIS_seqs_in_context}}

\code{\link{OEIS_seqs_adjacent}}
}
