% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_sequence.R
\name{OEIS_date}
\alias{OEIS_date}
\title{OEIS sequence date from \code{seq_df} data.frame}
\usage{
OEIS_date(seq_df)
}
\arguments{
\item{seq_df}{A data.frame with sequence content}
}
\value{
A \code{Date} object with the OEIS sequence date or NULL if it is not
  found
}
\description{
OEIS sequence date from \code{seq_df} data.frame
}
\examples{
id <- "A221863"
test_seq_html <- OEIS_xml2(id)
seq_df <- OEIS_df(test_seq_html)
OEIS_date(seq_df)
}
