% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_description.R
\name{OEIS_description}
\alias{OEIS_description}
\title{Get OEIS sequence description from \strong{internal format}}
\usage{
OEIS_description(internal_format)
}
\arguments{
\item{internal_format}{A character with the sequence internal format.}
}
\value{
A string with the OEIS sequence description
}
\description{
Sequence description gives a brief definition of the sequence. In the
description, a(n) usually denotes the n-th term of the sequence, and n is a
typical subscript. In some cases however n denotes a typical term in the
sequence.
}
\examples{
\dontrun{
id <- "A000056"
internal_format <- OEIS_internal_format(id)
OEIS_description(internal_format)
}
}
\seealso{
\code{\link{OEIS_internal_format}}

\code{\link{OEIS_sequence}}
}
