% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_url.R
\name{OEIS_description}
\alias{OEIS_description}
\title{OEIS_description:}
\usage{
OEIS_description(seq_xml)
}
\arguments{
\item{seq_xml}{A xml_document from an OEIS sequence}
}
\description{
OEIS_description:
}
\examples{
id <- "A000056"
test_seq_xml <- OEIS_html(id)
OEIS_description(test_seq_xml)
}
