% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS.R_functions.R
\name{OEIS_description}
\alias{OEIS_description}
\title{OEIS_description: Get OEIS sequence description from xml2 data}
\usage{
OEIS_description(seq_xml)
}
\arguments{
\item{seq_xml}{A xml_document from an OEIS sequence}
}
\value{
A string with the OEIS sequence description
}
\description{
OEIS_description: Get OEIS sequence description from xml2 data
}
\examples{
id <- "A000056"
test_seq_xml <- OEIS_xml2(id)
OEIS_description(test_seq_xml)
}
