% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_df.R
\name{OEIS_df}
\alias{OEIS_df}
\title{Get OEIS sequence information lines from XML document data}
\usage{
OEIS_df(seq_xml)
}
\arguments{
\item{seq_xml}{A class \code{OEIS_xml} document extracted from an OEIS
sequence.}
}
\value{
A \code{data.frame} with the sequence information lines.
}
\description{
This lines are useful to get all the sequence format data.
}
\examples{

id <- "A008406"
xml <- OEIS_xml(id)
df <- OEIS_df(xml)
df

}
\references{
\url{https://oeis.org/eishelp2.html#RS}{Explanation of Terms Used in Reply From}
}
\seealso{
\itemize{
\item \code{\link[=OEIS_xml]{OEIS_xml()}}
}

\itemize{
\item \code{\link[=OEIS_sequence]{OEIS_sequence()}}
}
}
