% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_df.R
\name{OEIS_df}
\alias{OEIS_df}
\title{Get OEIS sequence information lines from XML document data}
\usage{
OEIS_df(seq_xml)
}
\arguments{
\item{seq_xml}{A xml_document from an OEIS sequence}
}
\value{
A \code{data.frame} with sequence information lines.
}
\description{
This lines are useful to get all the sequence format data.
}
\references{
\url{https://oeis.org/eishelp2.html#RS}
}
\seealso{
\code{\link{OEIS_xml2}}
}
