% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_example.R
\name{OEIS_example}
\alias{OEIS_example}
\title{OEIS sequence example lines from sequence \code{internal_format}}
\usage{
OEIS_example(internal_format)
}
\arguments{
\item{internal_format}{A S3 object of the class \code{OEIS_internal} with the
sequence internal format.}
}
\value{
A character string with the OEIS sequence lines of examples or
  \code{NULL} if there are no examples.
}
\description{
These lines give expanded information or examples to illustrate the initial
terms of the sequence.
}
\examples{
\dontrun{
id <- "A105805"
internal_format <- OEIS_internal_format(id)
example <- OEIS_example(internal_format)
cat(example, sep = "\\n")
}
}
\seealso{
\code{\link{OEIS_description}}

\code{\link{OEIS_internal_format}}

\code{\link{OEIS_formula}}

\code{\link{OEIS_sequence}}
}
