% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_example.R
\name{OEIS_example}
\alias{OEIS_example}
\title{OEIS sequence example lines from sequence \code{internal_format}}
\usage{
OEIS_example(x)
}
\arguments{
\item{x}{Can be one of the following:
\itemize{
\item A character string with sequence \code{ID}.
\item A \code{OEIS_internal} S3 class object, with sequence internal format.
\item A S3 classes object \code{OEIS_xml} and \code{xml_document}, with
sequence content from OEIS web.
\item An object of \code{OEIS_sequence} class, that contains all sequence
related data.
}}
}
\value{
A character string with the OEIS sequence lines of examples or
  \code{NULL} where no examples are available.
}
\description{
These lines give expanded information or examples to illustrate the initial
terms of the sequence.
}
\examples{
id <- "A105805"
internal_format <- OEIS_internal_format(id)
example <- OEIS_example(internal_format)
cat(example, sep = "\n")

}
\seealso{
* \code{\link{OEIS_description}}

* \code{\link{OEIS_author}}

* \code{\link{OEIS_internal_format}}

* \code{\link{OEIS_formula}}

* \code{\link{OEIS_sequence}}

* \code{\link{OEIS_xml}}
}
