% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_formula.R
\name{OEIS_formula}
\alias{OEIS_formula}
\title{OEIS sequence formula lines from sequence \code{internal_format}}
\usage{
OEIS_formula(internal_format)
}
\arguments{
\item{internal_format}{A S3 object of the class \code{OEIS_internal} with the
sequence internal format.}
}
\value{
A character string with the OEIS sequence formula lines or
  \code{NULL} if there are no formulae.
}
\description{
These lines give formulae, recurrences, generating functions, etc. for the
sequence.
\strong{a(n)} usually denotes the n-th term of the sequence, and \strong{n}
is a typical subscript.
}
\examples{
\dontrun{
id <- "A105805"
internal_format <- OEIS_internal_format(id)
formula <- OEIS_formula(internal_format)
cat(formula, sep = "\\n")
}
}
\seealso{
\code{\link{OEIS_description}}

\code{\link{OEIS_internal_format}}

\code{\link{OEIS_sequence}}

\code{\link{OEIS_example}}
}
