% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_ggplot.R
\name{OEIS_ggplot}
\alias{OEIS_ggplot}
\title{OEIS_ggplot
Plot OEIS sequence with ggplot}
\usage{
OEIS_ggplot(x)
}
\arguments{
\item{x}{An object of class \code{OEIS_sequence}}
}
\value{
A graphic object of the classes `gg` and `ggplot`
}
\description{
OEIS_ggplot
Plot OEIS sequence with ggplot
}
\examples{
\dontrun{
# A123456: Ludwig van Beethoven, Bagatelle No. 25, Für Elise.
id <- "A123456"
x <- OEIS_sequence(id)
OEIS_ggplot(x)
}
}
\seealso{
\code{\link{OEIS_bibtex}}
}
