% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_keywords.R
\name{OEIS_keywords}
\alias{OEIS_keywords}
\title{OEIS sequence keywords from sequence \code{internal_format}}
\usage{
OEIS_keywords(internal_format)
}
\arguments{
\item{internal_format}{A character with the sequence internal format.}
}
\value{
A character vector with the OEIS sequence keywords
}
\description{
These lines give keywords describing the sequence.
The actual keywords in use can be found at:
\href{https://oeis.org/eishelp2.html}{Explanation of Terms Used}
}
\examples{
\dontrun{
id <- "A000012"
internal_format <- OEIS_internal_format(id)
OEIS_keywords(internal_format)
}
}
\seealso{
\code{\link{OEIS_author}}

\code{\link{OEIS_sequence}}

\code{\link{OEIS_internal_format}}
}
