% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_sequence.R
\name{OEIS_offset}
\alias{OEIS_offset}
\title{Get OEIS sequence offset from xml2 data}
\usage{
OEIS_offset(seq_xml)
}
\arguments{
\item{seq_xml}{A xml_document from an OEIS sequence}
}
\value{
A numeric vector with the OEIS sequence distinct offsets
}
\description{
Get OEIS sequence offset from xml2 data
}
\examples{
id <- "A000023"
test_seq_html <- OEIS_xml2(id)
OEIS_offset(test_seq_html)
}
