% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_crossrefs.R
\name{OEIS_seqs_adjacent}
\alias{OEIS_seqs_adjacent}
\title{OEIS Adjacent sequences from XML document}
\usage{
OEIS_seqs_adjacent(seq_xml)
}
\arguments{
\item{seq_xml}{A xml_document from an OEIS sequence}
}
\value{
A character vector OEIS sequences in context IDs
}
\description{
Adjacent sequences IDs in OEIS. These are the three sequences whose A-numbers
are immediately before and after the A-number of the sequence.
}
\examples{
id <- "A000112"
test_seq_html <- OEIS_xml2(id)
OEIS_seqs_adjacent(test_seq_html)
}
\seealso{
\code{\link{OEIS_seqs_in_context}}

\code{\link{OEIS_crossrefs}}

\code{\link{OEIS_cf}}

\code{\link{OEIS_xml2}}
}
