% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_sequence.R
\name{OEIS_sequence}
\alias{OEIS_sequence}
\title{OEIS_sequence: Class constructor}
\usage{
OEIS_sequence(ID)
}
\arguments{
\item{ID}{A string with the OEIS sequence ID}
}
\value{
An object of the class \code{OEIS_sequence}
}
\description{
OEIS_sequence: Class constructor
}
\examples{
id <- "A003456"
x <- OEIS_sequence(id)
x
}
