% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_sequence.R
\name{OEIS_sequence}
\alias{OEIS_sequence}
\title{Class constructor for \code{OEIS_sequence}}
\usage{
OEIS_sequence(ID)
}
\arguments{
\item{ID}{A string with the OEIS sequence \code{ID} Number
The A-number or sequence \code{ID} is the absolute catalogue number of the
sequence. It consists of "A" followed by 6 digits}
}
\value{
An object of the class \code{OEIS_sequence}
}
\description{
Class constructor for \code{OEIS_sequence}
}
\examples{
id <- "A003456"
A003456 <- OEIS_sequence(id)
A003456
}
\seealso{
\code{\link{OEIS_bfile}}
}
