% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_terms.R
\name{OEIS_terms}
\alias{OEIS_terms}
\title{Get OEIS sequence terms from XML document.}
\usage{
OEIS_terms(seq_xml)
}
\arguments{
\item{seq_xml}{A xml_document from an OEIS sequence}
}
\value{
A character list with the OEIS sequence terms.
}
\description{
This is the data included in the sequence. These lines give the beginning of
the sequence. Ideally the entry gives enough terms to fill about three lines
on the screen. The terms must be integers. If the terms are fractions, then
the numerators and denominators appear as separate sequences, labeled with
the Keyword "frac", and with links connecting the two sequences. Only
sequences that are well-defined and of general interest are included
}
\examples{
id <- "A000056"
test_seq_html <- OEIS_xml2(id)
OEIS_terms(test_seq_html)
}
\seealso{
\code{\link{OEIS_xml2}}
}
