% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_total_sequences.R
\name{OEIS_total_sequences}
\alias{OEIS_total_sequences}
\title{Total Number of sequences in OEIS}
\usage{
OEIS_total_sequences()
}
\value{
An integer with the total number of sequences in the Encyclopedia.
}
\description{
Total Number of sequences in OEIS
}
\examples{
#  The MIT License (MIT)()
}
