% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_url.R
\name{OEIS_url}
\alias{OEIS_url}
\title{OEIS_url:  Get sequence url}
\usage{
OEIS_url(ID)
}
\arguments{
\item{ID}{A string with the OEIS sequence ID}
}
\value{
A string with the sequence url
}
\description{
OEIS_url:  Get sequence url
}
\examples{
OEIS_url("A000055")
}
