% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_sequence.R
\name{OEIS_web_url}
\alias{OEIS_web_url}
\title{Get OEIS web main page url}
\usage{
OEIS_web_url()
}
\value{
A string with the OEIS web url
}
\description{
Get OEIS web main page url
}
\examples{
OEIS_web_url()
}
