% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_sequence.R
\name{OEIS_xml2}
\alias{OEIS_xml2}
\title{Get sequence content from OEIS web}
\usage{
OEIS_xml2(ID)
}
\arguments{
\item{ID}{A string with the OEIS sequence \code{ID} Number
The A-number or sequence \code{ID} is the absolute catalogue number of the
sequence. It consists of A followed by 6 digits}
}
\value{
\code{xml_document} from OEIS web with the sequence content
}
\description{
Get sequence content from OEIS web
}
\examples{
 id <- "A000156"
 test_seq_html <- OEIS_xml2(id)
}
