% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OEIS_xml2.R
\name{OEIS_xml2}
\alias{OEIS_xml2}
\title{Get sequence content from OEIS web as an XML document}
\usage{
OEIS_xml2(ID)
}
\arguments{
\item{ID}{A string with the OEIS sequence \code{ID} Number
The A-number or sequence \code{ID} is the absolute catalogue number of the
sequence. It consists of "A" followed by 6 digits}
}
\value{
\code{xml_document} from OEIS web with the sequence content
}
\description{
The XML document contains all the information from the OEIS web related to
the sequence.
}
\examples{
 id <- "A000156"
 test_seq_html <- OEIS_xml2(id)
}
\seealso{
\code{\link{OEIS_url}}

\code{\link{OEIS_check}}
}
